#295 [GCalendar] In share calendar view,delete icon should be pointer and show the confirm message

+++ src/main/webapp/js/ShareCalendar.js
@@ -83,17 +83,19 @@
 
             events:{
             	"click;.sharedDel":function(event){
-            		data = {};
-            		data.calendarId = view.calendarId;
-					var $btn = $(event.currentTarget);
-					var $div = $btn.closest(".ShareCalendar-item");
-					data.ruleId = $div.attr("data-id"); 
-            		app.googleApi.deleteSharedCalendar(data).done(function (extradata) {
-						setTimeout((function() {
-							$(document).trigger("DO_REFRESH_CALENDAR");
-						}), 3000); 
-						view.close();
-                    });
+                    if(confirm("Are you sure to delete it?")){
+                        data = {};
+                        data.calendarId = view.calendarId;
+                        var $btn = $(event.currentTarget);
+                        var $div = $btn.closest(".ShareCalendar-item");
+                        data.ruleId = $div.attr("data-id");
+                        app.googleApi.deleteSharedCalendar(data).done(function (extradata) {
+                            setTimeout((function() {
+                                $(document).trigger("DO_REFRESH_CALENDAR");
+                            }), 3000);
+                            view.close();
+                        });
+                    }
             	},
                 "btap; .ShareCalendarBtn":function () {
                     var view = this;
